CLI-508, CLI-509: Config loading (#528)
* attempt

* fix analytics attempt

* move config loading to main again

* move config loading and analytics init to command.go and add config as prerun fields

* add analytics non user for when config cannot be loaded

* fix to pass all unit tests

* fix cliname setting for config

* fix cli name bug

* remove local command merge resultion mistakes

* revert anayltics mock change

* fix mock helper

* go.mod and go.sum

* fix merge issues

* clean up merge

* clean up merge

+++ internal/pkg/cmd/prerunner.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"context"
+	v3 "github.com/confluentinc/cli/internal/pkg/config/v3"
 	"os"
 	"strings"
 
@@ -14,7 +15,6 @@ import (
 	"github.com/confluentinc/cli/internal/pkg/analytics"
 	pauth "github.com/confluentinc/cli/internal/pkg/auth"
 	v2 "github.com/confluentinc/cli/internal/pkg/config/v2"
-	v3 "github.com/confluentinc/cli/internal/pkg/config/v3"
 	"github.com/confluentinc/cli/internal/pkg/errors"
 	"github.com/confluentinc/cli/internal/pkg/log"
 	"github.com/confluentinc/cli/internal/pkg/update"
@@ -31,6 +31,8 @@ type PreRunner interface {
 
 // PreRun is the standard PreRunner implementation
 type PreRun struct {
+	Config             *v3.Config
+	ConfigLoadingError error
 	UpdateClient       update.Client
 	CLIName            string
 	Logger             *log.Logger
@@ -64,13 +66,14 @@ type HasAPIKeyCLICommand struct {
 func (a *AuthenticatedCLICommand) AuthToken() string {
 	return a.State.AuthToken
 }
+
 func (a *AuthenticatedCLICommand) EnvironmentId() string {
 	return a.State.Auth.Account.Id
 }
 
-func NewAuthenticatedCLICommand(command *cobra.Command, cfg *v3.Config, prerunner PreRunner) *AuthenticatedCLICommand {
+func NewAuthenticatedCLICommand(command *cobra.Command, prerunner PreRunner) *AuthenticatedCLICommand {
 	cmd := &AuthenticatedCLICommand{
-		CLICommand: NewCLICommand(command, cfg, prerunner),
+		CLICommand: NewCLICommand(command, prerunner),
 		Context:    nil,
 		State:      nil,
 	}
@@ -79,9 +82,9 @@ func NewAuthenticatedCLICommand(command *cobra.Command, cfg *v3.Config, prerunne
 	return cmd
 }
 
-func NewAuthenticatedWithMDSCLICommand(command *cobra.Command, cfg *v3.Config, prerunner PreRunner) *AuthenticatedCLICommand {
+func NewAuthenticatedWithMDSCLICommand(command *cobra.Command, prerunner PreRunner) *AuthenticatedCLICommand {
 	cmd := &AuthenticatedCLICommand{
-		CLICommand: NewCLICommand(command, cfg, prerunner),
+		CLICommand: NewCLICommand(command, prerunner),
 		Context:    nil,
 		State:      nil,
 	}
@@ -90,9 +93,9 @@ func NewAuthenticatedWithMDSCLICommand(command *cobra.Command, cfg *v3.Config, p
 	return cmd
 }
 
-func NewHasAPIKeyCLICommand(command *cobra.Command, cfg *v3.Config, prerunner PreRunner) *HasAPIKeyCLICommand {
+func NewHasAPIKeyCLICommand(command *cobra.Command, prerunner PreRunner) *HasAPIKeyCLICommand {
 	cmd := &HasAPIKeyCLICommand{
-		CLICommand: NewCLICommand(command, cfg, prerunner),
+		CLICommand: NewCLICommand(command, prerunner),
 		Context:    nil,
 	}
 	command.PersistentPreRunE = prerunner.HasAPIKey(cmd)
@@ -100,16 +103,16 @@ func NewHasAPIKeyCLICommand(command *cobra.Command, cfg *v3.Config, prerunner Pr
 	return cmd
 }
 
-func NewAnonymousCLICommand(command *cobra.Command, cfg *v3.Config, prerunner PreRunner) *CLICommand {
-	cmd := NewCLICommand(command, cfg, prerunner)
+func NewAnonymousCLICommand(command *cobra.Command, prerunner PreRunner) *CLICommand {
+	cmd := NewCLICommand(command, prerunner)
 	command.PersistentPreRunE = prerunner.Anonymous(cmd)
 	cmd.Command = command
 	return cmd
 }
 
-func NewCLICommand(command *cobra.Command, cfg *v3.Config, prerunner PreRunner) *CLICommand {
+func NewCLICommand(command *cobra.Command, prerunner PreRunner) *CLICommand {
 	return &CLICommand{
-		Config:    NewDynamicConfig(cfg, nil, nil),
+		Config: &DynamicConfig{},
 		Command:   command,
 		prerunner: prerunner,
 	}
@@ -128,43 +131,58 @@ func (h *HasAPIKeyCLICommand) AddCommand(command *cobra.Command) {
 // Anonymous provides PreRun operations for commands that may be run without a logged-in user
 func (r *PreRun) Anonymous(command *CLICommand) func(cmd *cobra.Command, args []string) error {
 	return func(cmd *cobra.Command, args []string) error {
+		r.Analytics.TrackCommand(cmd, args)
+		command.Config.Config = r.Config
 		command.Version = r.Version
 		command.Config.Resolver = r.FlagResolver
 		if err := log.SetLoggingVerbosity(cmd, r.Logger); err != nil {
 			return errors.HandleCommon(err, cmd)
 		}
-
+		r.Logger.Flush()
 		if err := r.notifyIfUpdateAvailable(cmd, r.CLIName, command.Version.Version); err != nil {
 			return errors.HandleCommon(err, cmd)
 		}
 		r.warnIfConfluentLocal(cmd)
-
-		ctx, err := command.Config.Context(cmd)
-		if err != nil {
-			return err
-		}
-		err = r.validateToken(cmd, ctx)
-		switch err.(type) {
-		case *ccloud.ExpiredTokenError:
-			err := ctx.DeleteUserAuth()
+		if r.Config != nil {
+			ctx, err := command.Config.Context(cmd)
 			if err != nil {
 				return err
 			}
-			ErrPrintln(cmd, "Your token has expired. You are now logged out.")
-			analyticsError := r.Analytics.SessionTimedOut()
-			if analyticsError != nil {
-				r.Logger.Debug(analyticsError.Error())
+			err = r.validateToken(cmd, ctx)
+			switch err.(type) {
+			case *ccloud.ExpiredTokenError:
+				err := ctx.DeleteUserAuth()
+				if err != nil {
+					return err
+				}
+				ErrPrintln(cmd, "Your token has expired. You are now logged out.")
+				analyticsError := r.Analytics.SessionTimedOut()
+				if analyticsError != nil {
+					r.Logger.Debug(analyticsError.Error())
+				}
+			}
+		} else {
+			if isAuthOrConfigCommands(cmd) {
+				return errors.HandleCommon(r.ConfigLoadingError, cmd)
 			}
 		}
-		r.Analytics.TrackCommand(cmd, args)
 		return nil
 	}
+}
+
+func isAuthOrConfigCommands(cmd *cobra.Command) bool {
+	return strings.Contains(cmd.CommandPath(), "login") ||
+		strings.Contains(cmd.CommandPath(), "logout") ||
+		strings.Contains(cmd.CommandPath(), "config")
 }
 
 // Authenticated provides PreRun operations for commands that require a logged-in Confluent Cloud user.
 func (r *PreRun) Authenticated(command *AuthenticatedCLICommand) func(cmd *cobra.Command, args []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		err := r.Anonymous(command.CLICommand)(cmd, args)
+		if r.Config == nil {
+			return errors.HandleCommon(r.ConfigLoadingError, cmd)
+		}
 		if err != nil {
 			return errors.HandleCommon(err, cmd)
 		}
@@ -195,6 +213,9 @@ func (r *PreRun) AuthenticatedWithMDS(command *AuthenticatedCLICommand) func(cmd
 		if err != nil {
 			return errors.HandleCommon(err, cmd)
 		}
+		if r.Config == nil {
+			return errors.HandleCommon(r.ConfigLoadingError, cmd)
+		}
 		err = r.setClients(command)
 		if err != nil {
 			return errors.HandleCommon(err, cmd)
@@ -221,6 +242,9 @@ func (r *PreRun) HasAPIKey(command *HasAPIKeyCLICommand) func(cmd *cobra.Command
 		err := r.Anonymous(command.CLICommand)(cmd, args)
 		if err != nil {
 			return errors.HandleCommon(err, cmd)
+		}
+		if r.Config == nil {
+			return errors.HandleCommon(r.ConfigLoadingError, cmd)
 		}
 		ctx, err := command.Config.Context(cmd)
 		if err != nil {
